implement user auth, registration and permission

+++ base/views.py
@@ -1,15 +1,63 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from base.forms import RoomForm
 from .models import Room, Topic
 
-# Create your views here.
 
-# rooms = [
-#     {'id': 1, 'name': 'Lets learn Nodejs!'},
-#     {'id': 2, 'name': 'UI/UX with Gee Lawrence'},
-#     {'id': 3, 'name': 'DevOps to the Moon'}
-# ]
+def loginPage(request):
+    
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('base:home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
+        
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('base:home')
+        else:
+            messages.error(request, "Invalid Credentials")
+        
+    context = {'page':page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    
+    logout(request)
+    return redirect('base:home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('base:home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+    
+    context = {'form':form}
+    return render(request, 'base/login_register.html', context)
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ""
@@ -31,6 +79,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='base:login')
 def createRoom(request):
     form = RoomForm()
     
@@ -43,9 +92,13 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='base:login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('Permission Denied!!!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -57,8 +110,12 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url='base:login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('Permission Denied!!!')
     
     if request.method == 'POST':
         room.delete()
